Implement SIP2-35: Added domain model
Added the domain model that will be used to bridge SIP2 and FOLIO. The model is based on the SIP2 specification, but implemented as Java POJOs. The intent is that a SIP2 parser will fill in the details of requests and the FOLIO communication layer will fill in the details of the response messages.

Currently, each message is a discrete object. There is some overlap that could be abstracted. For example, the `CheckoutResponse` and `RenewResponse` classes are identical, except for documentation. As we integrate the code, we may wish to refactor these classes so that common properties exist once.

Added checkstyle to the build to keep the code consistent.

Removed the `docker/docker-entrypoint.sh` script. It is no longer required.