Towards SIP2-42: Create a SIP2 to POJO parser
Created a simple parser that parses a SIP2 request message and returns a `Message` structure that contains error detection and the domain request object. This object will be used by the code that communicates with FOLIO.

The parser was integrated with the verticle in a very basic way. More work will be needed to harden this integration.

Not all the request commands are parsed. This work will follow on via another PR. However, there should be enough here to start integration with FOLIO.