Added more error detection tests
Added a method for generating the checksum of a message. We should probably move this to a utility class as soon as we have more tests that will need to send SIP2 formatted messages.

Added a few tests with error detection enabled using the new checksum method. We still likely need more negative tests here.