Freemarker template updates
Refactor the library to use a field name approach which takes some of the SIP details out of the response templates and moves them to the library. Commands that use the same field can just reuse the library macro. Some fields are required for certain commands and optional for other commands, some of the existing macros will need to be updated to handle this as more response templates are implemented.