Implement SIP2-11: add login via FOLIO
Added the ability to login using the SIP2 Login message and translating that to FOLIO then returning the Login Response message with the results of the FOLIO login.

Also aded some basics for Checkin wrt SIP2-32.

Some refactoring occurred in order to support the asynchronous nature of vertx's web client. This has propagated into the implementation, via vertx `Future` objects, but it is not clear to me how we can avoid this at this time.