SIP2-51: fix screen message and print line
Changed screen message to be a `List<String>` and the template library to use the list macro when creating the SIR message. Changed print line to be a `List<String>` as well, however while we use a list macro to create the SIP message, we also respect the max print width that will potentially split a string in the list into multiple print line fields.