Make the screen message field repeatable
In this case "repeatable" means that if the length of the line is longer than some max length, we split it and add another field with the same command code to the existing message. Since screen message is a list, we already have multiple fields being sent, so this adds additional message breakdown.