Correct the "language" field in the lib template
For some reason, a delimiter was added to the `language` field in the library template. This is a fixed length field with no command code, so there is no reason for a delimiter. In fact, this would be a SIP2 protocol error by any client reading a message with the `language` field present. This delimiter was removed and, as a result, a buggy unit test was corrected.