Use OffsetDateTime instead of ZonedDateTime
When dealing with date/time objects in SIP, we mostly will use the local time of the SC, which will likely be in the tenant time zone specified in FOLIO. In this case, we only need the time zone to obtain the offset to apply on dates returned to the SC. The offset may vary depending on the date/time, e.g. a return date might be after daylight saving time starts, so the offset will be different than the transaction date, which will still be in standard time.

The reverse applies on dates to be delivered to FOLIO from the SC, like the checkin date. Here we will need to convert the, most likely, local time zone (again retrieved from FOLIO) to UTC before sending the date to FOLIO.

None of these conversions require a ZonedDateTime object. We can pass time zone info to the SC in the SIP2 date/time format, however, this format requires military time zones, which are largely obsolete since they cannot express all time zones used word wide or minutes based offsets. While we do have conversion code, it is unlikely we'll need to use it.