SIP2-38: Add support for TLS (SSL) startup via vertx config
Vertx already supports SSL/TLS but we need to be able to configure it. Now we allow startup config (JSON) to be passed to the existing `NetServerOptions` that can contain SSL specific options or any net server option if some other setting might be necessary. When not specified, SSL/TLS support is not enabled.

Also changed `log4j2.xml` to accept a system supplied variable to set the log level at runtime. To enable, pass `-Dloglevel=debug` (or whatever level) to the VM. The default level is "info".

Finally, added some config and security documentation to the README.md file.