SIP2-37: return proper patron status and screen messages
If the patron is invalid or blocked in some way, we will now return the patron status to reflect this and generic screen messages:

```
Your library card number cannot be located.  Please see a staff member for assistance.
There are unresolved issues with your account.  Please see a staff member for assistance.
```