SIP2-58: Add the title identifier (AJ) to the checkin response
The self service kiosk being tested in Chalmers has a requirement that the title identifier (AJ) must be present in the SIP checkin response. However, this field is optional in SIP, so we ignored it to limit scope. This requirement was unknown and could not be known until actual integration testing.

We now add the title of the item into AJ field for the SIP checkin response. If title is not present, then we do not send it and the original error is shown in the kiosk. In this case, it is expected since there is likely a database problem here in that the title is not set in FOLIO.