SIP2-58: default to item identifier when the title is not available
In the rare case where FOLIO does not send a `title` field in the checkin-by-barcode response, we will, as recommended by Kiosk vendor Bibliotheca, use the item identifier (barcode) as the title. This will allow the Kiosk to provide some info for the patron to correlate the checkin response to an item in the case where many items are placed on the Kiosk.