SIP2-60: metrics for edge-sip2
Add support for Micrometer collection of metrics for delivery to a runtime defined backend. Micrometer claims to be the SLF4J for metrics, in that we code to a vendor agnostic API and rely on vendor specific bindings, that can be easily swapped in or out, to deliver metrics.

Also fixed a problem when bad data is received (the command was not being set to unknown).

Added a dependency analysis plugin for maven to better sort out dependencies (explicit vs implicit).

Added missing release maven plugin for when we do release.

Updated versions of dependencies. Note, we cannot update to Vert.x 3.7.1 unless one of these is done:
https://github.com/vert-x3/vertx-web/issues/1315
https://issues.folio.org/browse/SIP2-63

Added extensive documentation to the README for metrics setup and use.