SIP2-61: Enable PIN verification
We will now verify any command with a patron password sent if the tenant setting `patronPasswordVerificationRequired` is set to true. In this case, we attempt to log in to FOLIO. If we succeed, then the patron password is valid and the command can continue. If not, we shortcut the command and return a SIP2 message indicating command failure. If the setting is not true, then we do not attempt login and proceed with the command as usual.

SIP2-66 was also fixed here since end patron session sends a patron password.

SIP2-68 was also addressed in order to provide for a way to set the requirement for password verification.