SIP2-76: Fix unit test that uses fixed hours adjustment
A unit test started failing when daylight saving time ended in the
timezone used for a test since we were manually adjusting the clock by a
fixed offset to determine the expected date and time returned in a SIP
response. The offset is now 1 instead of 2, so the test will fail until
DST begins again in the timezone.

To correct this, since a master build shouldn't remain broken for
months, we'll just use java.time to compute the expected date and time
based on the test's timezone. This will handle DST shifts automatically.